Made simple screen
I made a simple screen, it was a little bit tricky to get working but it
works. I also fixed some reading issues in the main loop and have
removed the temporary stop.
I also added in case 0x02, which is like mov [reg], val.

+++ Emulator/src/include/emulator.h
@@ -28,7 +28,7 @@ typedef enum{
 	I4,
 	REGISTER_MAX
 } REGISTERS_T;
-uint16_t registers[REGISTER_MAX];
+uint32_t registers[REGISTER_MAX];
 
 typedef enum{
 	EQUAL_FLAG = 0,
